Feat: changing in all the app the long id to int id in player due is not supported by auto_increment

+++ src/main/java/cat/itacademyS5_01/game/service/GameService.java
@@ -12,7 +12,7 @@ public class GameService {
     @Autowired
     ReactiveMongoTemplate reactiveMongoTemplate;
 
-    public Mono<Game> findById(Long id) {
+    public Mono<Game> findById(int id) {
         return reactiveMongoTemplate.findById(id, Game.class);
     }
